Updated MVC stucture in files and folders
