+++ databases/orders.db
@@ -2,3 +2,4 @@
 {"ordernumber":"210f1ae3-b5e5-4833-8809-a49281b93f4d","user":"Gäst","eta":14,"placed":"2024-4-16 14:34:39","cart":[{"quantity":2,"item":{"id":3,"title":"Cappuccino","desc":"Bryggd på månadens bönor.","price":49}},{"quantity":3,"item":{"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39}}],"_id":"SR1mXkuBKQy5ey9A"}
 {"ordernumber":"57657e17-a58c-42e8-a29a-85f5c77337bd","user":"Gäst","eta":20,"placed":"2024-4-16 14:35:59","cart":[{"quantity":2,"item":{"id":3,"title":"Cappuccino","desc":"Bryggd på månadens bönor.","price":49}},{"quantity":3,"item":{"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39}}],"_id":"eF1GYNlkHe3k1URs"}
 {"ordernumber":"a92d96e0-e61b-49ca-9f30-52764e95ec08","user":"Gäst","eta":14,"placed":"2024-4-16 15:57:45","cart":[{"quantity":2,"item":{"id":3,"title":"Cappuccino","desc":"Bryggd på månadens bönor.","price":49}},{"quantity":3,"item":{"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39}}],"_id":"mai3WcBIyJfs2hTb"}
+{"ordernumber":"fbc44fd0-51e7-4667-8c24-814a10d5da3e","user":"c9c7771b-1475-4864-8456-1935f3175ee0","eta":18,"placed":"2024-4-17 11:42:28","cart":[{"quantity":2,"item":{"id":3,"title":"Cappuccino","desc":"Bryggd på månadens bönor.","price":49}},{"quantity":3,"item":{"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39}}],"_id":"UZJDbYKCO9r1lUlE"}
