+++ databases/users.db
@@ -1,3 +1,4 @@
+{"username":"Jacob","email":"jag@my.mail","password":"hejehj","userid":"3ca9723e-4023-450e-a8fd-0cd3afedeb37","_id":"09ll0Mpgy4YGryGL"}
+{"username":"Rasmus","email":"nalle@my-mail","password":"hejdå","userid":"eb0e3688-7105-4c03-a682-37bcc41e9175","_id":"FLsiys3xXTcxJqBJ"}
 {"username":"Egil","email":"egil@egilsson.se","password":"egil","userid":"476d0f6c-bf79-4a7d-bc43-5510b5160dc3","_id":"HDLRZM4KwWGJN7nc"}
 {"username":"Egilina","email":"egilina@egildottir.se","password":"egilina","userid":"bcbe96fe-56f4-490a-9b69-caead40c7502","_id":"SKyeIpH1ssPq3MWo"}
-{"username":"Jacob","email":"jag@my.mail","password":"hejehj","userid":"3ca9723e-4023-450e-a8fd-0cd3afedeb37","_id":"09ll0Mpgy4YGryGL"}
